+++ app/src/main/java/com/example/dndhandbook/presentation/screen/monsterList/MonsterListUIState.kt
@@ -8,5 +8,5 @@ data class MonsterListUIState(
     val monsterList: DefaultList = DefaultList(),
     val filterList: DefaultList = DefaultList(),
     val error: String = "",
-    val collectionId: Int? = null,
+    val collectionId: Long? = null,
 )
